requirements.txt
numpy==2.0.1
PyYAML==6.0.1
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4